pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Игорь Рудич <rudi4.clarinet@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
mysql-connector-python = "^8.4.0"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.0.0"
marshmallow = "^3.21.2"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
bootstrap-flask = "^2.4.0"
flask-wtf = "^1.2.1"
flask-admin = "^1.6.1"
flask-login = "^0.6.3"
itsdangerous = "^2.2.0"
flask-mail = "^0.10.0"
email-validator = "^2.2.0"
wtforms = "^3.1.2"
wtforms-sqlalchemy = "^0.4.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
